Read/write/DMs policy for global client
